refactoring
- change the type of AssignStack::particpated
- drop features:
  - progress_Bucket
  - luby_blocking
- convert the following flags into features:
  - clause_reduction
  - reason_side_rewarding
  - use_luby
- drop flags
  - a_adaptive
  - a_luby
  - a_reduce
  - a_rsr
- (propagate) change the order of sentences
- (propagate) prefer `get_unchecked``

+++ src/assign/mod.rs
@@ -74,7 +74,7 @@ pub struct Var {
     /// reverse conversion to index. Note `VarId` must be `usize`.
     pub index: VarId,
     /// the number of participation in conflict analysis
-    participated: f64,
+    participated: u32,
     /// a dynamic evaluation criterion like EVSIDS or ACID.
     reward: f64,
     /// the number of conflicts at which this var was assigned an rewarded lastly.
